pyproject.toml
[tool.poetry]
name = "hotel-rental"
version = "0.13.0"
description = ""
authors = ["Ilia Samodin <broobooot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
sqlalchemy = {extras = ["all"], version = "^2.0.29"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
black = "^24.3.0"
autoflake = "^2.3.1"
isort = "^5.13.2"
sqlalchemy-utils = "^0.41.2"
pytest = "^8.1.1"
sqlparse = "^0.4.4"
asyncio = "^3.4.3"
toml = "^0.10.2"
pytest-asyncio = "^0.23.6"
icecream = "^2.1.3"
email-validator = "^2.1.1"
phonenumbers = "^8.13.35"
bcrypt = "^4.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
celery = "^5.4.0"
flower = "^2.0.1"
psycopg2-binary = "^2.9.10"
sqladmin = "^0.20.1"
ruff = "^0.8.3"
pre-commit = "^4.0.1"
flake8 = "^7.1.1"
loguru = "^0.7.3"
orjson = "^3.10.12"
sentry-sdk = {extras = ["fastapi"], version = "^2.19.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
